Flight Review local server (#27)
* install flight review with logloader

* fix newlines end of file

* fix flight review install

* use repo install scripts for logloader, polaris, and rid-transmitter

* add manufacturer code and serial number to setup script

* README

* use systemctl --user to stop services before reinstall

* use sudo -v

* add function to keep sudo credentials alive, sometimes credentials will expire and require user to re-enter the password

* measure script execution time

* flight review allow multiple web socket origins

* add user network control logic to main install script, move hotspot-control to main install script

* function to sanitize y/n input during prompting, no input (enter) uses default value

* initilize sudo before process

* cleanup

* logloader prompt ask for local server

* adjust cleanup() position

* added flight review startup script, use hostname in bash for server

* fix hostname, try 5 times for mavsdk dpkg

* file permissions

* fix typos

* fix sudoers

* add http://

* add timestamp to each line in output.txt

* jetson doesn't support strftime

* silence pushd

* fix domain_name

* echo install jetpack for timing in output.txt

* fix files as sudo

* fix flashing for ardupilot

* give all scripts XDG paths, put mavlink-router main.conf at ~/.local/share/mavlink-router for non sudo editing

* major cleanup

* restart mavlink router

* reset flight controller after flash

* unify mavlink-router service between jetson and pi

* remove jetson from sudoers, allow ALL

* rename function, remove polaris binary

* remove /lib/systemd/system/.service

* sanitize input for RID mfg code and serial

* syntax error

* fix polaris service removal, account for name change

* fix flight-review install on pi

+++ setup/install_mavsdk.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
-
-sudo true
 source $(dirname $BASH_SOURCE)/functions.sh
 
 # Check if we are on 20.04 or 22.04
@@ -39,13 +37,22 @@ elif [ "$(lsb_release -cs)" = "jammy" ]; then
 		((attempt_num++))
 	done
 
-	if [ "$success" = true ]; then
-		echo "Downloading completed successfully."
-		echo "Installing $file_name"
-		sudo dpkg -i "$file_name"
-		sudo rm "$file_name"
-		sudo ldconfig
-	else
+if [ "$success" = true ]; then
+	echo "Downloading completed successfully."
+	echo "Installing $file_name"
+
+	for attempt in {1..5}; do
+		sudo dpkg -i "$file_name" && break || sleep 5
+	done
+
+	if [ $attempt -eq 5 ]; then
+		echo "Failed to install $file_name after 5 attempts."
+		exit 1
+	fi
+
+	sudo rm "$file_name"
+	sudo ldconfig
+else
 		echo "Failed to download the file after $max_attempts attempts."
 	fi
 else
